% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-adam_pasi.R
\docType{data}
\name{adam_pasi}
\alias{adam_pasi}
\title{Data: ADaM PASI}
\format{
A list with to ADaM datasets:
\describe{
  \item{ADSL}{subject level dataset}
  \item{ADPASI}{PASI records, follows basic data structure (BDS)}
}
}
\source{
Derived from
  Wonderful Wednesdays GitHub data repository (2021-04-14).
}
\usage{
adam_pasi
}
\description{
The Psoriasis Area and Severity Index (PASI) is a continuous measurement
of severity, the lower, the better the patient condition. The
derived dataset represents patient level follow-up during 8 visits over a
year of monitoring.
}
\examples{
str(adam_pasi, 1)

}
\keyword{datasets}
